Fixed some gcc warnings...

+++ protocols/facebook.cpp
@@ -65,7 +65,7 @@ std::string FacebookProtocol::replace(std::string &str, const char *string_to_re
 	// Find the first string to replace
 	int index = str.find(string_to_replace);
 	// while there is one
-	while(index != std::string::npos)
+	while(index != (int) std::string::npos)
 	{
 		// Replace it
 		str.replace(index, strlen(string_to_replace), new_string);
